moved tour logic to /lib/tour
