+++ src/components/Tour.vue
@@ -1,67 +1,74 @@
 <template>
-    <div class="tour">
-        <v-tour
-            name="nimiq-tour"
-            :steps="Object.values(steps).map((s) => s.tooltip)"
-        >
-            <template slot-scope="tour">
-                <transition name="fade">
-                    <v-step
-                        class="tooltip-step"
-                        v-if="tour.steps[tour.currentStep]"
-                        :key="tour.currentStep"
-                        :step="tour.steps[tour.currentStep]"
-                        :previous-step="tour.previousStep"
-                        :next-step="tour.goToNextStep"
-                        :stop="tour.stop"
-                        :skip="tour.skip"
-                        :is-first="tour.isFirst"
-                        :is-last="tour.isLast"
-                        :labels="tour.labels"
-                    >
-                        <div slot="content" class="content">
-                            <p
-                                v-for="(content, i) in tour.steps[tour.currentStep].content"
-                                :key="i"
-                                v-html="$t(content)"
-                            ></p>
-                            <!-- TODO REMOVE ME -->
-                            <div class="remove_me" v-if="currentStep === 1" @click="simulate()">
-                                Simulate Receive NIM
-                            </div>
+  <div class="tour">
+    <v-tour
+      name="nimiq-tour"
+      :steps="Object.values(steps).map((s) => s.tooltip)"
+    >
+        <template slot-scope="tour">
+            <transition name="fade">
+                <v-step
+                    class="tooltip-step"
+                    v-if="tour.steps[tour.currentStep]"
+                    :key="tour.currentStep"
+                    :step="tour.steps[tour.currentStep]"
+                    :previous-step="tour.previousStep"
+                    :next-step="tour.goToNextStep"
+                    :stop="tour.stop"
+                    :skip="tour.skip"
+                    :is-first="tour.isFirst"
+                    :is-last="tour.isLast"
+                    :labels="tour.labels"
+                >
+                    <div slot="content" class="content">
+                        <p v-for="(content, i) in tour.steps[tour.currentStep].content" :key="i" v-html="$t(content)">
+                        </p>
+                        <!-- TODO REMOVE ME -->
+                        <div class="remove_me" v-if="currentStep === 1" @click="simulate()">
+                            Simulate Receive NIM
                         </div>
-                        <div slot="actions" class="actions">
-                            <button @click="tour.previousStep" v-if="!isMobile">
-                                {{ $t("Previous step") }}
+                    </div>
+                    <div slot="actions">
+                        <div
+                            v-if="tour.steps[tour.currentStep].button || isFullDesktop"
+                            class="actions"
+                        >
+                            <button @click="goToPrevStep()" v-if="currentStep > 0 && isFullDesktop" class="left">
+                                <TourPreviousLeftArrowIcon />
+                                {{ $t("Previous") }}
                             </button>
-                            <button class="right" @click="() => tour.steps[tour.currentStep].button.fn()"
-                                v-if="tour.steps[tour.currentStep].button">
+                            <button
+                                class="right" v-if="tour.steps[tour.currentStep].button && !isLoading"
+                                @click="() => tour.steps[tour.currentStep].button.fn()"
+                            >
                                 {{ $t(tour.steps[tour.currentStep].button.text) }}
                             </button>
-                            <button class="right" @click="tour.goToNextStep"
-                                v-else-if="!isMobile">
-                                {{ $t("Next step") }}
+                            <button
+                                v-else-if="isFullDesktop && !disableNextStep && !isLoading"
+                                class="right"
+                                @click="goToNextStep()"
+                            >
+                                <span>{{ $t("Next") }}</span>
+                            </button>
+                            <button v-if="isLoading && isFullDesktop" class="circle-spinner right">
+                                <CircleSpinner />
                             </button>
                         </div>
-                    </v-step>
-                </transition>
-            </template>
-        </v-tour>
-        <transition name="fade">
-            <div class="tour-control-bar" v-if="isMobile">
+                    </div>
+                </v-step>
+            </transition>
+        </template>
+    </v-tour>
+    <transition name="fade">
+            <div class="tour-control-bar" v-if="isMobile || isTablet">
                 <button @click="endTour()">
                     {{ $t("End Tour") }}
                 </button>
-                <span class="progress">
-                    {{ currentStep + 1 }} / {{ nSteps }}
-                </span>
+                <span class="progress"> {{ currentStep + 1 }} / {{ nSteps }} </span>
                 <div class="arrows">
                     <button
+                        v-if="!isLoading"
                         class="prev"
-                        :class="{
-                            hidden: currentStep === 0,
-                        }"
-                        v-if="!loading"
+                        :class="{ hidden: currentStep === 0}"
                         @click="goToPrevStep()"
                         style="transform: rotate(180deg)"
                     >
@@ -69,18 +76,17 @@
                     </button>
                     <button
                         class="next"
-                        :class="{ loading }"
+                        :class="{ loading: isLoading }"
                         :disabled="disableNextStep"
-                        @click="!loading && goToNextStep()"
+                        @click="!isLoading && goToNextStep()"
                     >
-                        <CaretRightIcon v-if="!loading" />
+                        <CaretRightIcon v-if="!isLoading" />
                         <CircleSpinner v-else class="circle-spinner" />
                     </button>
                 </div>
             </div>
         </transition>
-
-    </div>
+  </div>
 </template>
 
 <script lang="ts">
@@ -98,9 +104,16 @@ import {
 } from '@vue/composition-api';
 import Vue from 'vue';
 import VueTour from 'vue-tour';
-import { MountedReturnFn, TourStep, TourStepIndex, TourSteps, useFakeTx, useTour } from '../composables/useTour';
 import { useWindowSize } from '../composables/useWindowSize';
+import {
+    getFakeTx,
+    getTour,
+    MountedReturnFn, TourStep,
+    TourStepIndex,
+    TourSteps,
+} from '../lib/tour';
 import CaretRightIcon from './icons/CaretRightIcon.vue';
+import TourPreviousLeftArrowIcon from './icons/TourPreviousLeftArrowIcon.vue';
 
 Vue.use(VueTour);
 
@@ -109,50 +122,49 @@ require('vue-tour/dist/vue-tour.css');
 export default defineComponent({
     name: 'tour',
     setup(props, context) {
-        // TODO Use isMobile
-        const { width } = useWindowSize();
+        const { isMobile, isTablet, isFullDesktop } = useWindowSize();
 
         const { state: $network } = useNetworkStore();
-        const disconnected = computed(
-            () => $network.consensus !== 'established',
-        );
+        const disconnected = computed(() => $network.consensus !== 'established');
 
         const { state: tourStore, setTour } = useAccountStore();
 
         let tour: VueTour.Tour | null = null;
-        const steps: TourSteps<any> = useTour(tourStore.tour, context) || {};
+        const steps: TourSteps<any> = getTour(tourStore.tour, context) || {};
 
         // Initial state
-        const loading = ref(true);
+        const isLoading = ref(true);
         const currentStep: Ref<TourStepIndex> = ref(0);
         const nSteps: Ref<number> = ref(0);
         const disableNextStep = ref(true);
 
         let unmounted: MountedReturnFn | void;
 
-        onMounted(() => {
-            tourSetup();
+        onMounted(async () => {
+            await tourSetup();
 
             // REMOVE ME
             const { removeTransactions } = useTransactionsStore();
-            removeTransactions([useFakeTx()]);
+            removeTransactions([getFakeTx()]);
         });
 
         async function tourSetup() {
             await context.root.$nextTick(); // to ensure the DOM is ready
 
-            const step = steps[currentStep.value];
+            const step = steps[currentStep.value]!;
 
             // Update state
             nSteps.value = Object.keys(steps).length;
-            disableNextStep.value = currentStep.value >= nSteps.value - 1
-                || !!step.ui.isNextStepDisabled;
+            disableNextStep.value = currentStep.value >= nSteps.value - 1 || !!step.ui.isNextStepDisabled;
 
             await step.lifecycle?.created?.({ goToNextStep, goingForward: true });
 
             _addAttributes(step.ui, currentStep.value);
 
-            unmounted = await step.lifecycle?.mounted?.({ goToNextStep, goingForward: true });
+            unmounted = await step.lifecycle?.mounted?.({
+                goToNextStep,
+                goingForward: true,
+            });
 
             if (context.root.$route.path !== step.path) {
                 context.root.$router.push(step.path);
@@ -162,15 +174,15 @@ export default defineComponent({
 
             tour = context.root.$tours['nimiq-tour'];
             tour.start(`${currentStep.value}`);
-            loading.value = false;
+            isLoading.value = false;
         }
 
         // Dont allow user to interact with the page while it is loading
         // But allow to end it
-        watch([loading, disconnected], () => {
+        watch([isLoading, disconnected], () => {
             const app = document.querySelector('#app main') as HTMLDivElement;
 
-            if (loading.value || disconnected.value) {
+            if (isLoading.value || disconnected.value) {
                 app.setAttribute('data-non-interactable', '');
             } else {
                 app.removeAttribute('data-non-interactable');
@@ -195,24 +207,25 @@ export default defineComponent({
         ) {
             const goingForward = futureStepIndex > currentStepIndex;
 
-            const { path: currentPath } = steps[currentStepIndex];
-            const { path: futurePath, ui: futureUI, lifecycle: futureLifecycle } = steps[futureStepIndex];
+            const { path: currentPath } = steps[currentStepIndex]!;
+            const { path: futurePath, ui: futureUI, lifecycle: futureLifecycle } = steps[futureStepIndex]!;
 
-            loading.value = true;
+            isLoading.value = true;
             tour!.stop();
+            await sleep(500);
 
-            await unmounted?.({ goingForward });
+            if (unmounted) {
+                await unmounted({ goingForward });
+            }
 
             await futureLifecycle?.created?.({ goToNextStep, goingForward });
 
-            if (futurePath !== currentPath && context.root.$route.fullPath !== futurePath) {
-                try {
-                    // Default prepare DOM
-                    context.root.$router.push(futurePath);
-                    await context.root.$nextTick();
-                } catch {
-                    // Ignore error
-                }
+            if (
+                futurePath !== currentPath
+                && context.root.$route.fullPath !== futurePath
+            ) {
+                context.root.$router.push(futurePath);
+                await context.root.$nextTick();
             }
 
             _addAttributes(futureUI, futureStepIndex);
@@ -231,15 +244,21 @@ export default defineComponent({
             // goingForward ? tour!.nextStep() : tour!.previousStep();
 
             // onMountedStep
-            loading.value = false;
+            isLoading.value = false;
             disableNextStep.value = futureStepIndex >= nSteps.value - 1 || !!futureUI.isNextStepDisabled;
 
-            unmounted = await futureLifecycle?.mounted?.({ goToNextStep, goingForward });
+            unmounted = await futureLifecycle?.mounted?.({
+                goToNextStep,
+                goingForward,
+            });
 
             currentStep.value = futureStepIndex;
         }
 
-        function _addAttributes(uiConfig: TourStep['ui'], stepIndex: TourStepIndex) {
+        function _addAttributes(
+            uiConfig: TourStep['ui'],
+            stepIndex: TourStepIndex,
+        ) {
             const fadedElements = uiConfig.fadedElements || [];
             const disabledElements = uiConfig.disabledElements || [];
 
@@ -258,13 +277,17 @@ export default defineComponent({
         }
 
         function _removeAttributes(stepIndex: TourStepIndex) {
-            document.querySelectorAll(`[data-non-interactable="${stepIndex}"]`).forEach((el) => {
-                el.removeAttribute('data-non-interactable');
-            });
+            document
+                .querySelectorAll(`[data-non-interactable="${stepIndex}"]`)
+                .forEach((el) => {
+                    el.removeAttribute('data-non-interactable');
+                });
 
-            document.querySelectorAll(`[data-opacified="${stepIndex}"]`).forEach((el) => {
-                el.removeAttribute('data-opacified');
-            });
+            document
+                .querySelectorAll(`[data-opacified="${stepIndex}"]`)
+                .forEach((el) => {
+                    el.removeAttribute('data-opacified');
+                });
         }
 
         async function endTour() {
@@ -284,18 +307,20 @@ export default defineComponent({
         // TODO REMOVE ME - Simulate tx
         function simulate() {
             const { addTransactions } = useTransactionsStore();
-            addTransactions([useFakeTx()]);
+            addTransactions([getFakeTx()]);
             goToNextStep();
         }
 
         return {
             steps,
-            isMobile: width.value < 700,
+            isMobile,
+            isTablet,
+            isFullDesktop,
 
             // control bar
             currentStep,
             nSteps,
-            loading: disconnected || loading,
+            isLoading: disconnected || isLoading,
             disableNextStep,
 
             // actions
@@ -309,6 +334,7 @@ export default defineComponent({
     },
     components: {
         CaretRightIcon,
+        TourPreviousLeftArrowIcon,
         CircleSpinner,
     },
 });
@@ -320,161 +346,186 @@ export default defineComponent({
 // [data-opaified="1"] although the CSS selector that we can use is [data-opacified]. @see _removeAttributes
 
 [data-opacified] {
-    filter: opacity(0.3);
+  filter: opacity(0.3);
 }
 
-[data-non-interactable], [data-non-interactable] * {
-    user-select: none !important;
-    pointer-events: none !important;
+[data-non-interactable],
+[data-non-interactable] * {
+  user-select: none !important;
+  pointer-events: none !important;
+}
+
+#app > *:not(.tour) {
+  cursor: not-allowed;
 }
 </style>
 
 <style lang="scss" scoped>
 .tour {
-    position: relative;
+  position: relative;
 
-    button {
-        width: min-content;
-        white-space: nowrap;
-        font-size: 16px;
-        padding: 0.8rem 1.6rem;
+  button {
+    width: min-content;
+    white-space: nowrap;
+    font-size: 16px;
+    padding: 0.8rem 1.6rem;
 
-        text-align: center;
-        background: #ffffff33; // TODO Maybe move this to a CSS variable (?)
-        color: var(--nimiq-white);
-        border: none;
-        outline: var(--nimiq-);
-        border-radius: 9999px;
+    text-align: center;
+    background: #ffffff33; // TODO Maybe move this to a CSS variable (?)
+    color: var(--nimiq-white);
+    border: none;
+    outline: var(--nimiq-);
+    border-radius: 9999px;
+    cursor: pointer;
+  }
+
+  .tooltip-step {
+    background: radial-gradient(
+      100% 100% at 100% 100%,
+      #265dd7 0%,
+      #0582ca 100%
+    );
+    box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.07),
+      0px 1.5px 3px rgba(0, 0, 0, 0.05),
+      0px 0.337011px 2px rgba(0, 0, 0, 0.0254662);
+    border-radius: 1rem;
+    width: clamp(200px, 255px, calc(100vw - 2rem));
+    padding: 2rem;
+
+    .content {
+      display: flex;
+      flex-direction: column;
+      gap: 1rem;
+
+      p {
+        margin: 0;
+        font-size: 15px;
+        line-height: 21px;
+        text-align: left;
+
+        br {
+          margin: 4rem 0;
+        }
+      }
     }
 
-    .tooltip-step {
-        background: radial-gradient(
-            100% 100% at 100% 100%,
-            #265dd7 0%,
-            #0582ca 100%
-        );
-        box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.07),
-            0px 1.5px 3px rgba(0, 0, 0, 0.05),
-            0px 0.337011px 2px rgba(0, 0, 0, 0.0254662);
-        border-radius: 1rem;
-        width: clamp(200px, 255px, calc(100vw - 2rem));
-        padding: 2rem;
+    .actions {
+      margin-top: 2rem;
+      display: flex;
 
-        .content {
-            display: flex;
-            flex-direction: column;
-            gap: 1rem;
+      button {
+        font-weight: 700;
+        font-size: 14px;
 
-            p {
-                margin: 0;
-                font-size: 15px;
-                line-height: 21px;
-                text-align: left;
+        &.left {
+          display: flex;
+          align-items: center;
+          opacity: 0.7;
+          background: none;
 
-                br {
-                    margin: 4rem 0;
-                }
-            }
+          svg {
+            margin-right: 0.75rem;
+          }
         }
 
-        .actions {
-            margin-top: 2rem;
-            display: flex;
-
-            button {
-                font-weight: 700;
-                font-size: 14px;
-
-                &.right {
-                    margin-left: auto;
-                }
-            }
-
+        &.right {
+          margin-left: auto;
         }
+
+        &.circle-spinner {
+          width: 28.78px;
+          height: 28.78px;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          padding: 0.5rem;
+
+          & ::v-deep svg path:nth-child(1) {
+            stroke: var(--nimiq-white);
+          }
+        }
+      }
     }
+  }
 }
 
 .tour-control-bar {
-    position: fixed;
-    bottom: 10px;
-    right: 0;
-    left: 0;
-    margin: 0 auto;
-    width: clamp(200px, 75vw, calc(100vw - 2rem));
-    padding: 1rem;
-    border-radius: 9999px;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    align-items: center;
-    z-index: 6;
+  position: fixed;
+  bottom: 10px;
+  right: 0;
+  left: 0;
+  margin: 0 auto;
+  width: clamp(200px, 75vw, calc(100vw - 2rem));
+  padding: 1rem;
+  border-radius: 9999px;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  align-items: center;
+  z-index: 6;
 
-    // TODO Cannot use CSS variables here
-    background: radial-gradient(
-        100% 100% at 100% 100%,
-        #265dd7 0%,
-        #0582ca 100%
-    );
+  // TODO Cannot use CSS variables here
+  background: radial-gradient(100% 100% at 100% 100%, #265dd7 0%, #0582ca 100%);
+
+  button {
+    padding: 1.4rem 1.6rem 1rem 1.6rem;
+    font-weight: 700;
+
+    &:disabled {
+      opacity: 0.5;
+    }
+
+    &.hidden {
+      display: none;
+    }
+  }
+
+  .arrows {
+    justify-self: end;
+    display: flex;
+    gap: 0.5rem;
 
     button {
-        padding: 1.4rem 1.6rem 1rem 1.6rem;
-        font-weight: 700;
+      height: 5rem;
+      width: 5rem;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding: 0;
 
-        &:disabled {
-            opacity: 0.5;
+      &.prev {
+        background: initial;
+      }
+
+      &.next {
+        padding-left: 0.75rem;
+
+        &.loading {
+          padding-left: 0;
+
+          & ::v-deep svg path:nth-child(1) {
+            stroke: var(--nimiq-white);
+          }
         }
-
-        &.hidden {
-            display: none;
-        }
+      }
     }
+  }
 
-    .arrows {
-        justify-self: end;
-        display: flex;
-        gap: 0.5rem;
-
-        button {
-            height: 5rem;
-            width: 5rem;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            padding: 0;
-
-            &.prev {
-                background: initial;
-            }
-
-            &.next {
-                padding-left: 0.75rem;
-
-                &.loading {
-                    padding-left: 0;
-
-                    & ::v-deep svg path:nth-child(1) {
-                        stroke: var(--nimiq-white);
-                    }
-                }
-            }
-        }
-    }
-
-    .progress {
-        opacity: 0.7;
-        color: var(--nimiq-white);
-        white-space: nowrap;
-        text-align: center;
-        font-size: 13px;
-    }
+  .progress {
+    opacity: 0.7;
+    color: var(--nimiq-white);
+    white-space: nowrap;
+    text-align: center;
+    font-size: 13px;
+  }
 }
 
 .remove_me {
-    background: var(--nimiq-green);
-    color: var(--nimiq-white);
-    border-radius: 9999px;
-    padding: 1rem 3rem;
-    font-size: 16px;
-    text-align: center;
-    cursor: pointer;
+  background: var(--nimiq-green);
+  color: var(--nimiq-white);
+  border-radius: 9999px;
+  padding: 1rem 3rem;
+  font-size: 16px;
+  text-align: center;
+  cursor: pointer;
 }
 </style>
