+++ src/components/icons/TourPreviousLeftArrowIcon.vue
@@ -1,0 +1,5 @@
+<template>
+    <svg width="12" height="9" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M4.5 8.125 1 4.625l3.5-3.5M2.25 4.593h8.499" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"/>
+    </svg>
+</template>
